tutorial/pyproject.toml
[tool.poetry]
name = "tutorial"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.0"
uvicorn = "^0.15.0"
SQLAlchemy = "^1.4.27"
requests = "^2.26.0"
pytest = "^6.2.5"
factory-boy = "^3.2.1"

[tool.poetry.dev-dependencies]
black = "^21.11b1"
isort = "^5.10.1"
pre-commit = "^2.15.0"
tox = "^3.24.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py36', 'py37']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
